requirements.txt
attrs
requests
tqdm
